update requirements; fix bugs

+++ solnml/components/transfer_learning/tlbo/acquisition_function/acquisition.py
@@ -73,7 +73,7 @@ class AbstractAcquisitionFunction(object, metaclass=abc.ABCMeta):
         acq = self._compute(X)
         if np.any(np.isnan(acq)):
             idx = np.where(np.isnan(acq))[0]
-            acq[idx, :] = -np.finfo(np.float).max
+            acq[idx, :] = -np.finfo(np.float32).max
         return acq
 
     @abc.abstractmethod
